add alias for experimental sycl namespace

+++ src/Graphics/D3D12/Utils/Device.hpp
@@ -86,7 +86,7 @@ private:
     // Device information (retrieved from adapter).
     std::string adapterName;
     uint32_t vendorId;
-    uint32_t adapterLuid;
+    uint64_t adapterLuid;
 };
 
 }}
